Rewrite wininet interop
